requirements.txt
streamlit>=1.28.0
boto3>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
botocore>=1.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
# fnllm-inspired dependencies for advanced features
aiolimiter>=1.1.0
tenacity>=8.5.0
typing-extensions>=4.0.0
# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
# CLI tools
click>=8.0.0